refactor: move MailClient, SmsClient, and TotpGenerator into utilServices
